Sweep dead code and duplication left by the extraction
Delete the never-called protocol helpers (observerSideToNamespace,
observerSideToNamespacePath, coerceObserverSide), the orphaned
*_HEARTBEAT event-name constants in config.js (event names live in the
library's protocol.ts), the utils/observerSide re-export shim (its two
importers now use the library entry point), and the dangling comments
left behind by deleted selectors. The initial-camera-selection effect,
copy-pasted verbatim between ObserverUI and the pilot's
CameraControlContainer, moves into a shared useInitialCameraSelection
hook.
