+++ react-frontend/src/features/observer-ui/ObserverUI.jsx
@@ -1,12 +1,7 @@
-import React, { useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
+import React from "react";
 import TopControlPanel from "./TopControlPanel";
 import { useObservedStation } from "../camera-controls/ObservedCameraProvider";
-import {
-  changeActiveCamera,
-  selectActiveCamera,
-  selectInitialCamHeartbeatData,
-} from "../camera-controls/cameraControlsSlice";
+import { useInitialCameraSelection } from "../camera-controls/useInitialCameraSelection";
 import CamHeartbeatListener from "../listeners/CamHeartbeatListener";
 import NewCameraCommandListener from "../listeners/NewCameraCommandListener";
 import RecorderHeartbeatListener from "../listeners/RecorderHeartbeatListener";
@@ -16,32 +11,8 @@ export default function ObserverUI({
   showFullCameraControls,
   setShowFullCameraControls,
 }) {
-  const dispatch = useDispatch();
-
   const station = useObservedStation();
-
-  const activeCamera = useSelector(selectActiveCamera);
-  const initialCamHeartbeat = useSelector(selectInitialCamHeartbeatData);
-
-  // use CAM_HEARTBEAT parameters only on initial app load to set activeCamera
-  // keep camera params in local state otherwise
-  useEffect(() => {
-    const setInitialCamera = () => {
-      dispatch(changeActiveCamera(initialCamHeartbeat));
-
-      // send camera change command to set available settings options
-      station.selectCamera(initialCamHeartbeat.camera, {
-        activeCamera: initialCamHeartbeat.camera,
-      });
-    };
-
-    // set initial camera state only if activeCamera is undefined
-    if (activeCamera === null) {
-      if (initialCamHeartbeat !== null) {
-        setInitialCamera();
-      }
-    }
-  }, [activeCamera, dispatch, station, initialCamHeartbeat]);
+  useInitialCameraSelection(station);
 
   return (
     <>
